k8s config file
this k8s config file indicates how to deploy an old application, we need to wait for config server started until the application will be deployed (because we need to read config information from config server). This config server may be a side car application who will work as a proxy between config server and the old application (if needed, the side car container can also be deployed in the same pod).